feat: optimized theme script + new script to open image dialog

+++ scripts/theme.sh
@@ -1,9 +1,18 @@
 WALLPAPER=$1
 CURR_WALLPAPER_PATH="$HOME/.cache/current_wallpaper.png"
 
+if [ ! -f "$WALLPAPER" ]; then
+  echo "$WALLPAPER doesn't exist"
+  exit 1
+fi
+
 # creatin symlink for current wallpaper for other services (like hyprlock)
 ln -sf "$(readlink -f $WALLPAPER)" "$CURR_WALLPAPER_PATH"
 
 awww img "$WALLPAPER"
 
-./wallbash.sh "$WALLPAPER"
+if [ ! -f "$WALLPAPER.dcol" ]; then
+  ~/dotfiles/scripts/wallbash.sh "$WALLPAPER"
+fi
+
+~/dotfiles/scripts/set-colors.sh "$WALLPAPER"
